pyproject.toml
[project]
name = "video-gen-llm-handler"
version = "0.1.0"
description = "FastAPI prompt splitter using Google Gemini on Fly.io"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.*",
    "uvicorn[standard]==0.30.*",
    "google-genai>=0.3.0",
    "pydantic==2.*",
    "httpx[http2]==0.27.*",
    "PyJWT==2.8.*",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
  "app",
  "README.md",
  "pyproject.toml",
]


